Move the function that applies deltas to a lookup-like structure to a common module (so GScan can re-use it too)
